Identifiable installer | Software REP
With a software ID Tag, it creates a better solution towards when the software package is installed on a machine the tag is installed along with it. The presence of the tag alleviates the problem of trying to “discover” what software has been installed on a machine and thereby improves the accuracy of identifying software assets automatically.